Fixing save log function
