+++ src/core/domain/codeReviewSettingsLog/contracts/codeReviewSettingsLog.service.contract.ts
@@ -1,3 +1,4 @@
+import { OrganizationAndTeamData } from '@/config/types/general/organizationAndTeamData';
 import { ICodeReviewSettingsLogRepository } from './codeReviewSettingsLog.repository.contract';
 
 export const CODE_REVIEW_SETTINGS_LOG_SERVICE_TOKEN = Symbol(
@@ -5,4 +6,10 @@ export const CODE_REVIEW_SETTINGS_LOG_SERVICE_TOKEN = Symbol(
 );
 
 export interface ICodeReviewSettingsLogService
-    extends ICodeReviewSettingsLogRepository {}
+    extends ICodeReviewSettingsLogRepository {
+    saveCodeReviewSettingsLog(
+        oldConfig: any,
+        newConfig: any,
+        organizationAndTeamData: OrganizationAndTeamData,
+    ): Promise<void>;
+}
